+++ zcash_client_sqlite/src/wallet/db.rs
@@ -109,6 +109,7 @@ CREATE TABLE "addresses" (
     transparent_child_index INTEGER,
     cached_transparent_receiver_address TEXT,
     exposed_at_height INTEGER,
+    receiver_flags INTEGER NOT NULL,
     FOREIGN KEY (account_id) REFERENCES accounts(id),
     CONSTRAINT diversification UNIQUE (account_id, key_scope, diversifier_index_be),
     CONSTRAINT transparent_index_consistency CHECK (