+++ zcash_client_sqlite/src/wallet/sapling.rs
@@ -367,6 +367,7 @@ pub(crate) fn ensure_address<
             diversifier_index,
             &ua,
             exposure_height,
+            false,
         )
         .map(Some)
     } else {